tests

+++ apps/chat/ui/build.ps1
@@ -7,25 +7,32 @@ $ErrorActionPreference = "Stop"
 . ../../../scripts/core.ps1
 
 
-Remove-Item ./dist -Recurse -Force -ErrorAction Ignore
+if (!$fast) {
+    Remove-Item ./dist -Recurse -Force -ErrorAction Ignore
 
-npx -y tailwindcss -i input.css -o dist/tailwind.css
+    npx -y tailwindcss -i input.css -o dist/tailwind.css
+}
 
 { trunk build --dist="dist/trunk" --public-url="./dist" } | Invoke-Block
 
 $path = "dist/trunk/index.html"
-(Get-Content $path) `
+$html = Get-Content $path
+$oldModule = ($html | Select-String -Pattern "import init from '/./dist/(.*?)'").Matches[0].Groups[1].Value
+$html `
     -replace "/./dist/", "./" `
+    -replace "import init from '([^']+)';init\(", "import init,{set_outer_messages_state,set_inner_messages_state} from './$oldModule';window.set_outer_messages_state=set_outer_messages_state;window.set_inner_messages_state=set_inner_messages_state;init(" `
     | Set-Content $path
 
 { rna build --bundle --minify --no-map --assetNames "[name]" dist/trunk/index.html --output dist/rna } | Invoke-Block
 
 $path = "dist/rna/index.html"
 $html = Get-Content $path
-$module = $html | Select-String -Pattern '<link rel="modulepreload" href="([^"]+)">' -AllMatches
-if ($module.Matches.Count -gt 0) {
-    $module.Matches | ForEach-Object { Remove-Item "dist/rna/$($_.Groups[1].Value)" -Force }
-    $html -replace ' *<link rel="modulepreload" href="([^"]+)">', "" | Set-Content $path
+
+$oldModule = $html | Select-String -Pattern '<link rel="modulepreload" href="([^"]+)">' -AllMatches
+if ($oldModule.Matches.Count -gt 0) {
+    $oldModule.Matches | ForEach-Object { Remove-Item "dist/rna/$($_.Groups[1].Value)" -Force }
+    $newModule = $html | Select-String -Pattern "import './(.*?)'"
+    $html -replace '<link rel="modulepreload" href="([^"]+)">', "<link rel=`"modulepreload`" href=`"$($newModule.Matches[0].Groups[1].Value)`">" | Set-Content $path
 }
 
 if (!$fast) {
